Implements logins and session handling.
- Added SessionsModel to Auth plugin.
- Changed Auth constants to global scope for use in the controllers.
- Added sources as a plugin extension type. Using sources, you can model 
a piece of data using a class, rather than implementing a set library 
and invoking a function. You can interact with the sources directly, as 
they are only ever imported. Sources will always need to be in the 
\Plugins\<name>\Src namespace.
- Made Account model for Auth plugin
- Added more config options for Auth. Also cleaned it up a bit.
- Altered bootstrap.php to find the user's IP address. We do this so we 
can better track session IP addresses for data collection and security 
purposes.
- Fixed forgotten-about method of deleting data from a table by 
implementing a delete method for the parent model. Only accepts in a 
where clause.
- Updated Cookie and Session to allow deletion of sessions and cookies. 
Can do it specifically or all at once by referring to destroy() and 
destroy_all() in each class.